# krt: allow lookup by Index, replace Namespace logic with Index