--- a/pkg/kube/krt/fetch.go
+++ b/pkg/kube/krt/fetch.go
@@ -40,10 +40,16 @@ func Fetch[T any](ctx HandlerContext, cc Collection[T], opts ...FetchOption) []T
 
 	// Now we can do the real fetching
 	var res []T
-	for _, i := range c.List(d.filter.namespace) {
+	var list []T
+	if d.filter.listFromIndex != nil {
+		list = d.filter.listFromIndex().([]T)
+	} else {
+		list = c.List()
+	}
+	for _, i := range list {
 		i := i
 		o := c.augment(i)
-		if d.filter.Matches(o) {
+		if d.filter.Matches(o, true) {
 			res = append(res, i)
 		}
 	}