--- a/pkg/kube/krt/join.go
+++ b/pkg/kube/krt/join.go
@@ -41,10 +41,10 @@ func (j *join[T]) GetKey(k Key[T]) *T {
 	return ptr.Of(j.merge(found))
 }
 
-func (j *join[T]) List(namespace string) []T {
+func (j *join[T]) List() []T {
 	res := map[Key[T]][]T{}
 	for _, c := range j.collections {
-		for _, i := range c.List(namespace) {
+		for _, i := range c.List() {
 			key := GetKey(i)
 			res[key] = append(res[key], i)
 		}