--- a/pkg/kube/krt/static.go
+++ b/pkg/kube/krt/static.go
@@ -14,7 +14,11 @@
 
 package krt
 
-import "istio.io/istio/pkg/maps"
+import (
+	"istio.io/istio/pkg/kube/controllers"
+	"istio.io/istio/pkg/maps"
+	"istio.io/istio/pkg/slices"
+)
 
 type staticList[T any] struct {
 	vals map[Key[T]]T
@@ -49,22 +53,12 @@ func (s *staticList[T]) augment(a any) any {
 	return a
 }
 
-func (s *staticList[T]) List(namespace string) []T {
-	if namespace == "" {
-		return maps.Values(s.vals)
-	}
-	var res []T
-	// Future improvement: shard outputs by namespace so we can query more efficiently
-	for _, v := range s.vals {
-		if getNamespace(v) == namespace {
-			res = append(res, v)
-		}
-	}
-	return res
+func (s *staticList[T]) List() []T {
+	return maps.Values(s.vals)
 }
 
 func (s *staticList[T]) Register(f func(o Event[T])) Syncer {
-	panic("StaticCollection does not support event handlers")
+	return registerHandlerAsBatched(s, f)
 }
 
 func (s *staticList[T]) Synced() Syncer {
@@ -72,7 +66,15 @@ func (s *staticList[T]) Synced() Syncer {
 }
 
 func (s *staticList[T]) RegisterBatch(f func(o []Event[T]), runExistingState bool) Syncer {
-	panic("StaticCollection does not support event handlers")
+	if runExistingState {
+		f(slices.Map(s.List(), func(e T) Event[T] {
+			return Event[T]{
+				New:   &e,
+				Event: controllers.EventAdd,
+			}
+		}))
+	}
+	return alwaysSynced{}
 }
 
 var _ internalCollection[any] = &staticList[any]{}