… pkg/kube/krt/index_test.go