This PR introduces a few changes:
* Allow Fetching by an Index. This is going to be used directly in some followup work
* Rewrite Namespace lookup/list to use an Index on Namespace. Before, the namespace was just a post-applied filtering anyways. We also didn't use any listing by a certain namespace anyways


A modest perf improvement (though the benchmark doesn't use it as effectively as future PR will, so its not the primary motivation):

```
name                   old time/op    new time/op    delta
Controllers/krt-16       14.7ms ± 4%    12.2ms ± 6%  -17.40%  (p=0.000 n=10+9)

name                   old alloc/op   new alloc/op   delta
Controllers/krt-16       16.0MB ± 0%    14.9MB ± 0%   -6.41%  (p=0.000 n=10+9)

name                   old allocs/op  new allocs/op  delta
Controllers/krt-16         125k ± 0%      122k ± 0%   -2.36%  (p=0.000 n=10+9)

```