@istio-testing 2024-04-09
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
lint_istio | d0dd75043240075e1c1e7a2f2a88cae689c2faed | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/50324/lint_istio/1777712187175866368) | true | `/test lint`
integ-security-multicluster_istio | d0dd75043240075e1c1e7a2f2a88cae689c2faed | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/50324/integ-security-multicluster_istio/1777712182146895872) | true | `/test integ-security-multicluster`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->